redesign post compile
I had some issues with postcompile namely:

1. Long compile times

The old design would use all dependencies that the crate that invoked
postcompile had. This was extremely bloated and resulted in long compile
times and resulted in cargo having difficulting caching the code.

2. Dependencies

There was no way to include dependencies that are not needed by the app
and only needed by the tests.

3. Tests

There was no way to execute the generated code (write a unit test) for
the code that was generated.